요일 괸련 클래스 테스트

+++ fittable-project/commons/src/main/java/net/fittable/domain/premises/Coordinate.java
@@ -14,7 +14,7 @@ public class Coordinate {
     private final double latitude;
     private final double longitude;
 
-    public Coordinate(float latitude, float longitude) {
+    public Coordinate(double latitude, double longitude) {
         this.latitude = latitude;
         this.longitude = longitude;
     }
